fix(lvda): correct rx_lvd_interrupts_t underlying type from uint16_t to uint8_t
Vector numbers 88 (LVD1) and 89 (LVD2) fit in uint8_t (0-255).
Per project coding standard, enums must use the smallest type that
fits all values. Table 15.5 (manual p.512) confirms both vectors.

+++ star-rx72n-firmware/libs/rx_hal/inc/rx72n_lvda_regs.h
@@ -501,9 +501,9 @@ typedef enum : uint8_t {
  *
  * @since Version 1.0.0
  */
-typedef enum : uint16_t {
-  k_lvd1_irq = 88, /**< LVD1 - Voltage monitor 1 interrupt */
-  k_lvd2_irq = 89, /**< LVD2 - Voltage monitor 2 interrupt */
+typedef enum : uint8_t {
+  k_lvd1_irq = 88, /**< LVD1 - Voltage monitor 1 interrupt (Table 15.5 p.512) */
+  k_lvd2_irq = 89, /**< LVD2 - Voltage monitor 2 interrupt (Table 15.5 p.512) */
 } rx_lvd_interrupts_t;
 
 /* =============================================================================
